Created VisitorBase for compilation visitors
