+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/compilation/VisitorTso.java
@@ -5,20 +5,16 @@ import com.dat3m.dartagnan.program.Register;
 import com.dat3m.dartagnan.program.event.Tag;
 import com.dat3m.dartagnan.program.event.EventFactory.X86;
 import com.dat3m.dartagnan.program.event.Tag.C11;
-import com.dat3m.dartagnan.program.event.arch.aarch64.StoreExclusive;
 import com.dat3m.dartagnan.program.event.arch.tso.Xchg;
 import com.dat3m.dartagnan.program.event.core.*;
 import com.dat3m.dartagnan.program.event.core.rmw.RMWStore;
 import com.dat3m.dartagnan.program.event.lang.catomic.*;
-import com.dat3m.dartagnan.program.event.lang.linux.*;
 import com.dat3m.dartagnan.program.event.lang.pthread.*;
 import com.dat3m.dartagnan.program.event.visitors.EventVisitor;
 import org.sosy_lab.common.configuration.Options;
 
-import java.util.Collections;
 import java.util.List;
 
-import static com.dat3m.dartagnan.configuration.Arch.*;
 import static com.dat3m.dartagnan.expression.op.COpBin.EQ;
 import static com.dat3m.dartagnan.expression.op.COpBin.NEQ;
 import static com.dat3m.dartagnan.program.event.EventFactory.eventSequence;
@@ -31,18 +27,11 @@ import static com.dat3m.dartagnan.program.event.EventFactory.newLocal;
 import static com.dat3m.dartagnan.program.event.EventFactory.newRMWLoad;
 import static com.dat3m.dartagnan.program.event.EventFactory.newRMWStore;
 import static com.dat3m.dartagnan.program.event.EventFactory.newStore;
-import static com.dat3m.dartagnan.program.processing.compilation.Compilation.commonVisitLock;
-import static com.dat3m.dartagnan.program.processing.compilation.Compilation.commonVisitUnlock;
 
 @Options
-public class VisitorTso implements EventVisitor<List<Event>> {
+public class VisitorTso extends VisitorBase implements EventVisitor<List<Event>> {
 
 	protected VisitorTso() {}
-
-	@Override
-	public List<Event> visitEvent(Event e) {
-		return Collections.singletonList(e);
-	};
 
 	@Override
 	public List<Event> visitCreate(Create e) {
@@ -64,13 +53,6 @@ public class VisitorTso implements EventVisitor<List<Event>> {
 	}
 
 	@Override
-	public List<Event> visitInitLock(InitLock e) {
-		return eventSequence(
-                newStore(e.getAddress(), e.getMemValue(), e.getMo())
-        );
-	}
-
-	@Override
 	public List<Event> visitJoin(Join e) {
         Register resultRegister = e.getResultRegister();
 		Load load = newLoad(resultRegister, e.getAddress(), e.getMo());
@@ -83,11 +65,6 @@ public class VisitorTso implements EventVisitor<List<Event>> {
 	}
 
 	@Override
-	public List<Event> visitLock(Lock e) {
-		return commonVisitLock(e);
-	}
-
-	@Override
 	public List<Event> visitStart(Start e) {
         Register resultRegister = e.getResultRegister();
 
@@ -96,56 +73,6 @@ public class VisitorTso implements EventVisitor<List<Event>> {
         		newJumpUnless(new Atom(resultRegister, EQ, IConst.ONE), e.getLabel())
         );
 	}
-
-	@Override
-	public List<Event> visitUnlock(Unlock e) {
-		return commonVisitUnlock(e);
-	}
-
-	@Override
-	public List<Event> visitStoreExclusive(StoreExclusive e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
-
-	@Override
-	public List<Event> visitRMWAbstract(RMWAbstract e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
-	
-	@Override
-	public List<Event> visitRMWAddUnless(RMWAddUnless e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
-	
-	@Override
-	public List<Event> visitRMWCmpXchg(RMWCmpXchg e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
-	
-	@Override
-	public List<Event> visitRMWFetchOp(RMWFetchOp e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
-	
-	@Override
-	public List<Event> visitRMWOp(RMWOp e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
-	
-	@Override
-	public List<Event> visitRMWOpAndTest(RMWOpAndTest e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
-	
-	@Override
-	public List<Event> visitRMWOpReturn(RMWOpReturn e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
-	
-	@Override
-	public List<Event> visitRMWXchg(RMWXchg e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
-	};
 
 	@Override
 	public List<Event> visitXchg(Xchg e) {
@@ -164,11 +91,6 @@ public class VisitorTso implements EventVisitor<List<Event>> {
                 store,
                 newLocal(resultRegister, dummyReg)
         );
-	}
-
-	@Override
-	public List<Event> visitAtomicAbstract(AtomicAbstract e) {
-		throw new IllegalArgumentException("Compilation to " + TSO + " is not supported for " + e.getClass().getName());
 	}
 
 	@Override
